+++ src/schemes/euler/eulerblackscholes.cpp
@@ -12,7 +12,7 @@
 
 
 
-State EulerBlackScholes::init_state(float S0,  std::optional<float> v0) const { 
+State EulerBlackScholes::init_state(double S0,  std::optional<double> v0) const { 
 
     State state{S0};
     return state;
@@ -24,11 +24,11 @@ State EulerBlackScholes::step(const State& state, int i, float dt, std::mt19937&
 
     if (dt <= 0) throw std::invalid_argument("EulerBlackScholes::step : dt must be stricltly positive");
 
-    std::normal_distribution<float> dist;
+    std::normal_distribution<double> dist;
     
-    float logS = std::log(state.spot());
-    float Z = dist(rng);
-    float logSt = logS + (model.mu - 0.5*model.sigma*model.sigma)*dt + model.sigma*Z * std::sqrt(dt);
+    double logS = std::log(state.spot());
+    double Z = dist(rng);
+    double logSt = logS + (model.mu - 0.5*model.sigma*model.sigma)*dt + model.sigma*Z * std::sqrt(dt);
 
     return State{std::exp(logSt)};
 }
